Rename flex_bison_parser.hpp to flex-bison-parser-constants.hpp. - The file just contains a constant value, not really a parser. - We have just added a cqasm-annotations-constants.hpp. - Also, all source files use dashes instead of underscores.
